+++ NEWS.md
@@ -3,6 +3,8 @@
 ## QFeatures 1.21.3
 
 - Add `readQFeaturesFromDIANN(multiplexing = 'dimethyl')` example.
+- New `replaceColnames()` function (see [Issue
+  258](https://github.com/rformassspectrometry/QFeatures/issues/258))
 
 ## QFeatures 1.21.2